move cert files outside src
