+++ packages/amplify-graphql-schema-generator/certs/aws-rds-global-bundle.pem
