+++ packages/amplify-graphql-schema-generator/certs/download-certs.sh
