--- packages/amplify-graphql-schema-generator/src/certs/index.ts
@@ -1,1 +1,0 @@
-export { getSSLConfig } from './cert';
