+++ packages/amplify-graphql-schema-generator/src/datasource-adapter/pg-datasource-adapter.ts
@@ -1,10 +1,10 @@
 import { knex } from 'knex';
-import { invokeSchemaInspectorLambda } from '../utils/vpc-helper';
 import ora from 'ora';
+import { invokeSchemaInspectorLambda } from '../utils/vpc-helper';
 import { Field, Index } from '../schema-representation';
+import { getSSLConfig } from '../utils';
 import { DataSourceAdapter } from './datasource-adapter';
 import { PostgresStringDataSourceAdapter, expectedColumns } from './pg-string-datasource-adapter';
-import { getSSLConfig } from '../certs';
 
 const spinner = ora();
