+++ packages/amplify-graphql-schema-generator/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './vpc-helper';
+export { getSSLConfig } from './cert';
