+++ packages/amplify-graphql-schema-generator/vpc-db-lambda/lib/aws-rds-global-bundle.pem
